fix build warnings: adding on top of abel-mak

+++ src/gamemanager.cpp
@@ -52,7 +52,7 @@ void Game_manager::generate_enemies()
 {
     static Rand count_rand {_Min_enemy_per_row, _Max_enemy_per_row};
     static Rand x_rand{_Enemy_first_allowed_column, _Enemy_last_allowed_column}; // these spaces will be occupied by the shape
-    constexpr int y = 0;
+    constexpr float y = 0;
 
     if (enemies.size() >= _Enemy_maximum_generate)
         return;
